[宽度处理]

+++ scaledbarcode.go
@@ -107,12 +107,12 @@ func scale2DCode(bc Barcode, width, height int) (Barcode, error) {
 func scale1DCode(bc Barcode, width, height int) (Barcode, error) {
 	orgBounds := bc.Bounds()
 	orgWidth := orgBounds.Max.X - orgBounds.Min.X
-	factor := int(float64(width) / float64(orgWidth))
+	factor := float64(width) / float64(orgWidth)
 
 	if factor <= 0 {
 		return nil, fmt.Errorf("can not scale barcode to an image smaller than %dx1", orgWidth)
 	}
-	width = orgWidth * factor
+	width = int(float64(orgWidth) * factor)
 	offsetX := 0
 
 	wrap := func(x, y int) color.Color {
